add authorization to users

+++ src/users/router.ts
@@ -1,11 +1,12 @@
 import { createUserSchema, updateUserSchema } from "./schema";
 import { HTTPException } from "hono/http-exception";
-import { Hono } from "hono";
+import { isAuthedMiddleware } from "../middleware/isAuthed";
+import { newPublicRouter } from "../middleware/public";
 import { paramIdSchema } from "../utils/validation";
 import userService from "./service";
 import { zValidator } from "@hono/zod-validator";
 
-const usersRouter = new Hono();
+const usersRouter = newPublicRouter();
 
 usersRouter.post("/", zValidator("json", createUserSchema), async (context) => {
 	const data = context.req.valid("json");
@@ -15,10 +16,13 @@ usersRouter.post("/", zValidator("json", createUserSchema), async (context) => {
 
 usersRouter.put(
 	"/:id",
+	isAuthedMiddleware,
 	zValidator("param", paramIdSchema),
 	zValidator("json", updateUserSchema),
 	async (context) => {
 		const { id } = context.req.valid("param");
+		const userId = context.get("userId");
+		if (id !== userId) throw new HTTPException(401, { message: "unauthorized" });
 		const data = context.req.valid("json");
 		const user = await userService.updateUser(id, data);
 		if (!user) throw new HTTPException(404, { message: "User does not exist" });
@@ -26,11 +30,18 @@ usersRouter.put(
 	},
 );
 
-usersRouter.delete("/:id", zValidator("param", paramIdSchema), async (context) => {
-	const { id } = context.req.valid("param");
-	const user = await userService.deleteUser(id);
-	if (!user) throw new HTTPException(404, { message: "User does not exist" });
-	return context.json(user, 200);
-});
+usersRouter.delete(
+	"/:id",
+	isAuthedMiddleware,
+	zValidator("param", paramIdSchema),
+	async (context) => {
+		const { id } = context.req.valid("param");
+		const userId = context.get("userId");
+		if (id !== userId) throw new HTTPException(401, { message: "unauthorized" });
+		const user = await userService.deleteUser(id);
+		if (!user) throw new HTTPException(404, { message: "User does not exist" });
+		return context.json(user, 200);
+	},
+);
 
 export default usersRouter;
